Rename to storybook-addon-knobs.d.ts (+1 squashed commit) Squashed commits: [4d5a906] Move types to root
